requirements.txt
Django
djangorestframework
psycopg2-binary
factory_boy
pyyaml
